Add v1beta1 deprovisioning consolidation testing
